add graduated feature

+++ database/seeders/BloodTableSeeder.php
@@ -8,20 +8,19 @@ use Illuminate\Support\Facades\DB;
 
 class BloodTableSeeder extends Seeder
 {
-    /**
-     * Run the database seeds.
-     *
-     * @return void
-     */
     public function run()
     {
-        
-        DB::table('blood__types')->truncate();
+
+        // if blood__types table has at least one data truncate if not do nothing
+        if (DB::table('blood__types')->first()) {
+            DB::table('blood__types')->truncate();
+        }
 
         $bgs = ['O-','O+', 'A+', 'A-', 'B+', 'B-', 'AB+' , 'AB-'];
 
         foreach ($bgs as $bg) {
             Blood_Type::Create(['name' => $bg]);
         }
+
     }
 }
